requirements.txt
speedtest-cli==2.1.3
requests==2.32.5